backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.0",
    "google-genai>=1.64.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.2.1",
    "sajupy>=0.2.0",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]
